backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
mysqlclient==2.2.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
PyJWT==2.11.0
